# spec/openapi.bundle.yaml
openapi: 3.0.3
info:
  title: Permission Slip API
  version: 0.1.0
  description: |
    Permission Slip is a centralized middle-man service that sits between AI agents and external 
    APIs. Agents submit pre-defined **actions** for human approval, and Permission Slip executes 
    them using the user's stored credentials. Agents never see user credentials and can never 
    make arbitrary API calls.

    **Base URL**: `https://app.permissionslip.dev`

    ## Key Concepts

    - **Action**: Pre-defined, structured request with a type, validated parameters, and execution logic — the core primitive
    - **Agent**: Automated system with a cryptographic key pair that submits actions for approval
    - **Approver**: Human authorized to approve or deny agent actions
    - **Connector**: Integration with an external service (e.g., Gmail, Stripe) that defines available actions
    - **Credential**: Encrypted API key or OAuth token for an external service, stored in Permission Slip's vault
    - **Registration**: One-time setup where agent proves identity via public key
    - **Approval Request**: Async request for permission to execute a specific action
    - **Token**: Single-use credential issued after one-off approval, used to execute the approved action
    - **Standing Approval**: Pre-authorized, constraint-scoped grant that lets an agent execute an action repeatedly without per-request approval

    ## Authentication

...

...
      See the `PermissionSlipSignature` security scheme.
...
servers:
...
paths:
...
  /invite/{invite_code}:
    post:
      operationId: registerAgent
      summary: Register Agent via Invite URL
      description: |
        Register a new agent with Permission Slip by POSTing to the invite URL.
        The invite code in the URL path identifies which user the agent is
        registering with — the agent does not need to know the user's username.

        The invite URL is generated by the user from the Permission Slip dashboard
        (`POST /v1/registration-invites`) and shared with the agent. The agent
        POSTs its public key to the URL to begin registration.

        ## Registration Flow

        1. **User generates invite**: Clicks "Add Agent" on dashboard, receives
           invite URL (e.g., `https://app.permissionslip.dev/invite/PS-R7K3-X9M4`)
        2. **User shares URL**: Gives the invite URL to the agent (paste in chat,
           config file, environment variable, etc.)
        3. **Agent POSTs to invite URL**: Sends public key and request ID
        4. **Permission Slip validates**: Verifies the invite code is valid,
           not expired, and not consumed — then creates a pending registration
        5. **Dashboard shows confirmation code**: User sees a 6-character
           confirmation code on their dashboard
        6. **User shares confirmation code**: Communicates the code to the agent
           (out-of-band)
        7. **Agent verifies**: Submits the code to
           `POST /v1/agents/{agent_id}/verify`
        8. **Registration complete**: Agent can now make approval requests

        ## Why no username?

        The invite code already maps to exactly one user in the database. When the
        user generates an invite, the backend stores their `user_id` alongside the
        hashed invite code. When the agent POSTs to the invite URL, the backend
        looks up the invite code and finds the user automatically. Requiring the
        agent to also specify the username would be redundant.

        ## Security

        - Request MUST include valid `X-Permission-Slip-Signature` header
        - The server auto-assigns an `agent_id` (bigint) returned in the response
        - Permission Slip enforces 5-minute timestamp window for signatures
        - Duplicate `request_id` values are rejected within timestamp window
        - Invite codes are single-use, time-limited (default 15 min), and locked
          after 5 failed attempts

        ## Idempotency

        If the same `request_id` is used multiple times (within 300 seconds),
        Permission Slip returns the original response if registration succeeded,
        or rejects with `duplicate_request_id` error.
      tags:
        - Registration
      security:
        - PermissionSlipSignature: []
      parameters:
        - $ref: '#/components/parameters/InviteCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
            examples:
              basic:
                summary: Basic registration
                value:
                  request_id: 6f1a7c30-9b2e-4d91-8c3f-2a4b6c7d8e9f
                  public_key: ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIFPZj8V7K2mT9Xw4nL5rQ1pY3vN8cM6hU0oB4eW2sR7k
              with_metadata:
                summary: Registration with agent metadata
                value:
                  request_id: 7a2b8d41-0c3e-5e02-9d4f-3b5c7d9e0f1a
                  public_key: ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIFPZj8V7K2mT9Xw4nL5rQ1pY3vN8cM6hU0oB4eW2sR7k
                  metadata:
                    name: My AI Assistant
                    version: 1.0.0
                    capabilities:
                      - email
                      - calendar
                      - payments
      responses:
        '200':
          description: Registration initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAgentResponse'
              example:
                agent_id: 42
                expires_at: '2026-02-11T13:25:00Z'
                verification_required: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid signature or invalid invite code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_signature:
                  summary: Signature verification failed
                  value:
                    error:
                      code: invalid_signature
                      message: Signature verification failed
                      retryable: false
                      trace_id: trace_xyz789
                invalid_invite_code:
                  summary: Invite code does not match
                  value:
                    error:
                      code: invalid_invite_code
                      message: Invite code is not valid
                      retryable: false
                      trace_id: trace_xyz789
        '404':
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invite_not_found
                  message: No invite found for this code
                  retryable: false
                  trace_id: trace_xyz789
        '409':
          description: Agent already registered or duplicate request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                already_registered:
                  summary: Agent already registered
                  value:
                    error:
                      code: agent_already_registered
                      message: Agent is already registered
                      retryable: false
                      details:
                        agent_id: 42
                      trace_id: trace_xyz789
                duplicate_request:
                  summary: Duplicate request ID
                  value:
                    error:
                      code: duplicate_request_id
                      message: Request ID already used
                      retryable: false
                      details:
                        request_id: 6f1a7c30-9b2e-4d91-8c3f-2a4b6c7d8e9f
                      trace_id: trace_xyz789
        '410':
          description: Invite expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invite_expired
                  message: Invite has expired — ask the user to generate a new one
                  retryable: false
                  trace_id: trace_xyz789
        '423':
          description: Invite locked after too many failed attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invite_locked
                  message: Invite locked after too many failed attempts — ask the user to generate a new one
                  retryable: false
                  details:
                    failed_attempts: 5
                    max_attempts: 5
                  trace_id: trace_xyz789
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /v1/agents:
    get:
      operationId: listAgents
      summary: List Agents
      description: |
        Returns agents owned by the authenticated user in any status
        (pending, registered, deactivated), ordered by creation time
        (newest first). Supports cursor-based pagination via the `limit`
        and `after` query parameters. Uses session authentication (Supabase JWT).
      tags:
        - Agents
      security:
        - SupabaseSession: []
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of agents to return (1–100, default 50).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          example: 20
        - name: after
          in: query
          required: false
          description: |
            Opaque pagination cursor. Pass the `next_cursor` value from a
            previous response to fetch the next page. Do not construct
            cursor values manually.
          schema:
            type: string
          example: 2026-02-11T13:15:00.123456Z,42
      responses:
        '200':
          description: Agents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
              example:
                data:
                  - agent_id: 42
                    status: registered
                    metadata:
                      name: My AI Assistant
                      version: 1.0.0
                    registered_at: '2026-02-11T13:20:15Z'
                    last_active_at: '2026-02-19T08:45:00Z'
                    request_count_30d: 42
                    created_at: '2026-02-11T13:15:00Z'
                  - agent_id: 43
                    status: pending
                    request_count_30d: 0
                    created_at: '2026-02-18T10:00:00Z'
                has_more: false
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invalid_request
                  message: limit must be a positive integer
                  retryable: false
                  trace_id: trace_a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Profile not found for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: profile_not_found
                  message: Profile not found
                  retryable: false
                  trace_id: trace_a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/agents/{agent_id}:
    get:
      operationId: getAgent
      summary: Get Agent
      description: |
        Returns a single agent by ID. The agent must belong to the authenticated user.
      tags:
        - Agents
      security:
        - SupabaseSession: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSummary'
              example:
                agent_id: 42
                status: registered
                metadata:
                  name: My AI Assistant
                  version: 1.0.0
                registered_at: '2026-02-11T13:20:15Z'
                last_active_at: '2026-02-19T08:45:00Z'
                created_at: '2026-02-11T13:15:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Agent not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: agent_not_found
                  message: Agent not found
                  retryable: false
                  trace_id: trace_a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    patch:
      operationId: updateAgent
      summary: Update Agent
      description: |
        Update an agent's metadata (e.g. name, description). The agent must belong
        to the authenticated user.
      tags:
        - Agents
      security:
        - SupabaseSession: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
            example:
              metadata:
                name: My Renamed Agent
      responses:
        '200':
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSummary'
              example:
                agent_id: 42
                status: registered
                metadata:
                  name: My Renamed Agent
                registered_at: '2026-02-11T13:20:15Z'
                last_active_at: '2026-02-19T08:45:00Z'
                created_at: '2026-02-11T13:15:00Z'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invalid_request
                  message: metadata is required
                  retryable: false
                  trace_id: trace_a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Agent not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: agent_not_found
                  message: Agent not found
                  retryable: false
                  trace_id: trace_a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/agents/{agent_id}/deactivate:
    post:
      operationId: deactivateAgent
      summary: Deactivate Agent
      description: |
        Deactivates an agent, preventing it from making further approval requests
        or executing actions. Any active standing approvals for the agent are
        automatically revoked. The agent must belong to the authenticated user.

        Returns 404 if the agent does not exist, does not belong to the
        authenticated user, or is already deactivated.
      tags:
        - Agents
      security:
        - SupabaseSession: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent deactivated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSummary'
              example:
                agent_id: 42
                status: deactivated
                metadata:
                  name: My AI Assistant
                registered_at: '2026-02-11T13:20:15Z'
                deactivated_at: '2026-02-18T09:30:00Z'
                last_active_at: '2026-02-18T08:15:00Z'
                created_at: '2026-02-11T13:15:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Agent not found, not owned by user, or already deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: agent_not_found
                  message: Agent not found
                  retryable: false
                  trace_id: trace_a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    PermissionSlipSignature:
      type: apiKey
      in: header
      name: X-Permission-Slip-Signature
      description: |
        Cryptographic signature proving agent identity. Required for all Permission Slip 
        endpoints except connector discovery (`GET /v1/connectors`).

        ## Header Format

        ```
        X-Permission-Slip-Signature: agent_id="<agent_id>", algorithm="<algorithm>", timestamp="<unix_timestamp>", signature="<base64url_signature>"
        ```

        ## Fields

        - `agent_id`: Auto-incrementing agent identifier (integer)
        - `algorithm`: Signature algorithm (`Ed25519` or `ECDSA-P256`)
        - `timestamp`: Unix timestamp in seconds (must be within 300 seconds of current time)
        - `signature`: Base64url-encoded signature (no padding)

        ## Canonical Request Format

        The signature is computed over a canonical representation:

        ```
        <HTTP_METHOD>\n
        <PATH>\n
        <QUERY_STRING>\n
        <TIMESTAMP>\n
        <BODY_HASH>
        ```

        ### Canonicalization Rules

        1. **HTTP_METHOD**: Uppercase HTTP verb (e.g., `POST`, `GET`)
        2. **PATH**: URL path component without query string (e.g., `/invite/PS-R7K3-X9M4`)
        3. **QUERY_STRING**: RFC 3986 percent-encoded, sorted parameters (empty string if no params)
           - Sort parameters by name (lexicographic order)
           - Use uppercase hex in percent-encoding (`%2F` not `%2f`)
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
      example:
        error:
          code: invalid_signature
          message: Signature verification failed
          retryable: false
          details:
            timestamp: '1770816000'
            expected_window: 300 seconds
          trace_id: trace_xyz789
...
  parameters:
...
    InviteCode:
      name: invite_code
      in: path
      required: true
      description: |
        Single-use invite code in the format `PS-XXXX-XXXX`. Generated by a user
        from the Permission Slip dashboard and shared with the agent. The code
        identifies which user the agent is registering with — the agent does not
        need to know the user's username.
      schema:
        type: string
        pattern: ^PS-[A-Z2-9]{4}-[A-Z2-9]{4}$
        maxLength: 12
      example: PS-R7K3-X9M4
...